Fix a crash on Connect, and four things found beside it
Two simulators on one Mac turn out to be two real peers. LocalNetwork-
Transport is Bonjour plus TCP with no BLE dependency and no simulator
guard, and it outranks BLE, so a simulator advertises _airlink._tcp,
browses, dials and runs the real handshake. That made an end-to-end test
possible for the first time on this machine, and it immediately found
things a thousand green tests had not.

THE CRASH. Pressing Connect segfaulted the app. `connect:endpointId:
timeoutMs:resolve:reject:` declared `timeoutMs` as `double`; codegen's
protocol declares it `NSInteger`, because the spec says Int32. That
compiles - a selector matches on its name - but arm64 passes integers in
x registers and doubles in v registers, so every argument after the
mismatch was read from the wrong place and `resolve` arrived as a garbage
pointer. Retaining it killed the process, in `objc_retain`, before any
of my code ran.

Clang warns about exactly this. scripts/build-ios.sh dropped every
warning on the floor to hide third-party pod noise, so the defect was
found by a crash instead. The script now never filters warnings in our
own sources, and takes a destination and configuration so it can build
for a device at all.

FOUR ROWS FOR ONE PHONE. A device publishes the same Bonjour service
type from BOTH local-network transports and browses with both, so every
device saw every other device four times, each with its own Connect
button. The advertisement token is what collapses them: one device
advertises one token across every transport at any instant. Strangers now
merge on it, endpoints anchor a row across a token rotation, and a device
that turns up later carrying an abandoned token is correctly a different
device. Merging strangers confers no trust - identity comes from the
handshake and the six digits, and a friend resolves by peer id and never
reaches this path.

"BLUETOOTH IS OFF" WHILE A PEER SITS BELOW IT. With Wi-Fi available the
local network finds people perfectly well, so "AirLink can't find friends
nearby" was contradicted by the friend in the list underneath. Same fact,
honest weight: a quieter banner that says Bluetooth reaches further and
works with no network at all.

`wifiOn` STOOD FOR SEVERAL TRANSPORTS AND TOOK WHICHEVER SPOKE LAST. Both
report at startup, so it landed on whichever finished second. It is an OR
now, asked of the client, which is the only thing that knows them all.

Verified by running it: one row, the handshake completes, both sides show
296 416, and confirming on both promotes the peer to a trusted friend.

Still broken and NOT fixed here: reconnecting to an already-paired peer
over the local network fails repeatedly with "Couldn't connect". The
failure is at least graceful - a calm banner and Try again, no crash -
but it is a real defect and it is written down rather than papered over.

+++ native/airlink-transport/ios/RCTNativeAirLinkTransport.mm
@@ -155,14 +155,27 @@ RCT_EXPORT_MODULE(NativeAirLinkTransport)
 
 #pragma mark - Links
 
+/**
+ * `timeoutMs` is NSInteger, and the type is load-bearing.
+ *
+ * The spec declares it `Int32`, so codegen puts `NSInteger` in the protocol.
+ * This was written as `double`, which compiles - a selector matches on its name
+ * - but does not work: on arm64 integers are passed in x registers and doubles
+ * in v registers, so the argument after the mismatch is read from the wrong
+ * place. `resolve` and `reject` arrived as garbage pointers and retaining one
+ * segfaulted the app the instant anybody pressed Connect.
+ *
+ * Clang does warn about this ("conflicting parameter types in implementation").
+ * The build script used to drop every warning on the floor; it no longer does.
+ */
 - (void)connect:(NSString *)transport
      endpointId:(NSString *)endpointId
-      timeoutMs:(double)timeoutMs
+      timeoutMs:(NSInteger)timeoutMs
         resolve:(RCTPromiseResolveBlock)resolve
          reject:(RCTPromiseRejectBlock)reject {
   [_bridge connect:transport
         endpointId:endpointId
-         timeoutMs:(NSInteger)timeoutMs
+         timeoutMs:timeoutMs
            resolve:^(NSString *linkId) { resolve(linkId); }
             reject:^(NSString *code, NSString *message) { reject(code, message, nil); }];
 }
